Move most code in system76 boards into system76/common
