+++ src/board/system76/common/include/board/board.h
@@ -1,7 +1,10 @@
 #ifndef _BOARD_BOARD_H
 #define _BOARD_BOARD_H
 
+#include <stdbool.h>
+
 void board_init(void);
 void board_event(void);
+void board_on_ac(bool ac);
 
 #endif // _BOARD_BOARD_H
