+++ src/board/system76/common/include/board/dgpu.h
@@ -1,6 +1,10 @@
 #ifndef _BOARD_DGPU_H
 #define _BOARD_DGPU_H
 
+#ifndef HAVE_DGPU
+    #define HAVE_DGPU 0
+#endif
+
 void dgpu_init(void);
 void dgpu_event(void);
