+++ src/board/system76/darp5/include/board/gpio.h
@@ -48,5 +48,6 @@ extern struct Gpio __code VA_EC_EN;
 extern struct Gpio __code VR_ON;
 extern struct Gpio __code WLAN_EN;
 extern struct Gpio __code WLAN_PWR_EN;
+#define HAVE_XLP_OUT 0
 
 #endif // _BOARD_GPIO_H
