+++ src/board/system76/gaze15/board.mk
@@ -25,11 +25,25 @@ CFLAGS+=-DI2C_SMBUS=I2C_4
 # Set touchpad PS2 bus
 CFLAGS+=-DPS2_TOUCHPAD=PS2_3
 
+# Custom fan curve
+CLFAGS+=-DBOARD_HEATUP=5
+CFLAGS+=-DBOARD_COOLDOWN=20
+CFLAGS+=-DBOARD_FAN_POINTS="\
+	FAN_POINT(65, 40), \
+	FAN_POINT(70, 60), \
+	FAN_POINT(75, 75), \
+	FAN_POINT(80, 90), \
+	FAN_POINT(85, 100) \
+"
+
 # Set smart charger parameters
 CFLAGS+=\
 	-DCHARGER_CHARGE_CURRENT=1536 \
 	-DCHARGER_CHARGE_VOLTAGE=16800 \
 	-DCHARGER_INPUT_CURRENT=9230
+
+# Enable DGPU support
+CFLAGS+=-DHAVE_DGPU=1
 
 # Set CPU power limits in watts
 CFLAGS+=\
