+++ src/board/system76/oryp6/board.mk
@@ -42,6 +42,9 @@ CFLAGS+=\
 	-DCHARGER_CHARGE_VOLTAGE=16800 \
 	-DCHARGER_INPUT_CURRENT=13050
 
+# Enable DGPU support
+CFLAGS+=-DHAVE_DGPU=1
+
 # Set CPU power limits in watts
 CFLAGS+=\
 	-DPOWER_LIMIT_AC=180 \
