--- src/board/system76/oryp6/include/board/dgpu.h
@@ -1,7 +1,0 @@
-#ifndef _BOARD_DGPU_H
-#define _BOARD_DGPU_H
-
-void dgpu_init(void);
-void dgpu_event(void);
-
-#endif // _BOARD_DGPU_H
