+++ src/board/system76/oryp6/include/board/gpio.h
@@ -25,7 +25,7 @@ extern struct Gpio __code EC_EN;
 extern struct Gpio __code EC_RSMRST_N;
 extern struct Gpio __code GC6_FB_EN;
 extern struct Gpio __code LED_ACIN;
-extern struct Gpio __code LED_AIRPLANE_N;
+#define HAVE_LED_AIRPLANE_N 0
 extern struct Gpio __code LED_BAT_CHG;
 extern struct Gpio __code LED_BAT_FULL;
 extern struct Gpio __code LED_PWR;
